LoxiGen: workaround a Tenjin bug
Because the variable context is shared between nested includes (aka dynamic
scope), and because Tenjin explicitly deletes variables from the context after
the include is finished, we have a problem where the outer template's context
variables are deleted by the inner template.

Fix it by subclassing tenjin.Engine and rewriting include. We can't delegate to
the original include because it uses stack frame magic.