pyloxi: support field length members
These are members whose value is the length of another member. This is used in
OpenFlow for variable length members in the middle of a message. Variable
length members at the end do not need a field length member because their
length is implicit. The field length members must come before, but not
necessarily immediately before, the field they refer to.

The only place this was used in OF 1.0 was in packet_out. Now that this is
gnerically supported we can remove the special case code for packet_out.