VlanVid: fix special case handling (untagged+present)
OF1.0 and OF1.3 differ in the magic values they assign to matching
untagged packets (OF1.0: 0xFFFF, OF1.3: 0x0000), and the presence
of the OFPVID_PRESENT flag (0x1000) in OF1.3. This change attempts
to handle these changes in a transparent way for the OF1.0 and OF1.3
line protocols.

Internally, the class uses the OF1.3 representation.