loci: only mask partially wildcarded fields
The of_match_values_mask function was consuming about 7% CPU time in my
benchmarks. For realistic flows most of the work it did was pointless, because
most fields are fully wildcarded and most of the rest are not wildcarded. Only
in the uncommon case of partially wildcarded fields does this function do
anything.

Fixed by calling a new of_memmask function only when parsing a masked OXM.