loci: lazily zero wire buffer
Previously we zeroed the entire wire buffer when it was allocated. This was
wasteful because often the message didn't grow to fill the entire allocated
size, and the memset was consuming 80% of CPU time during one of my benchmarks.

With this change, we only zero memory right before giving it to the user.
Additionally, of_octets_t fields like packet-in data aren't zeroed at all.