loci: add clean API to get OXM from a set-field action
Right now users of loci need to dig into the wire buffer and internal loci
functions to parse the OXM in a set-field action. This is because loci doesn't
have good support for subtyped fields (as opposed to list elements).

This commit doesn't add generic support for subtyped fields, but it does add a
reasonable API over the special case code.