wireshark: detect and skip obviously broken messages
When a TCP segment is lost Wireshark may attempt to start a PDU in the middle
of an OpenFlow message. This breaks framing and can make the rest of the
capture useless. These checks attempt to detect this and skip this segment,
hopefully restarting dissection at the beginning of the next message.