pyloxi: Connection class
This class creates a thread which continually parses OpenFlow messages off the
supplied socket and places them in a queue. The class has methods for reading messages
from the RX queue, sending messages, and higher level operations like request-response
and multipart transactions.

I'm putting this in Loxigen because it's so commonly used even though it goes
beyond the usual LOXI scope of serialization.