java: generate setters for unwriteable fields
https://github.com/floodlight/loxigen/pull/438 ran into a problem when we tried
to use a field as a discriminator when it was a normal data member in the
superclass. The java compiler complained that we hadn't provided an
implementation of the setter in the concrete class.

This change fixes the problem by creating setters for unwriteable fields which
throw an exception, like we do when the field isn't available in the OpenFlow
version.

This obviously adds the setters in more places than necessary. We could
possibly optimize this by checking each ancestor class to see if the field
changed writeability and only emit a setter in that special case.