Only generate virtual setters where necessary LOXI-76
The previous commit would add virtual Exception-raising setters
globally. This commit adds a method that specifically checks whether
a setter is necessary.

A setter must be generated if the field is writeable in this class
or any super class. Fields can turn non-writeable in a subclass
if a superclass data member is used as a discriminator.