BSC-11527 rework build: add Jenkinsfile, use MVN CI friendly versioning
This reworks the build process for Loxigen to run entirely in a custom
docker container, and with the Jenkins native docker agent support.

It also consolidates all of the build jobs for loxigen in a single
Jenkinsfile.

Lastly, it changes the versioning model for the MVN artificats deployed.

Master builds now become release (as opposed to SNAPSHOT builds); with
the patch version taken from the build number.

This means that dependent projects can reference a stable release
version that won't ever change; we don't need to manually bump the MVN
version.