LOXI-81: TransportPort - treat 0xFFFF as "NO_MASK"
TransportPort is a short (2-byte) field; is is read and written as 2
bytes. However, for normalization purposes, we were looking for a 4-byte
0xFFFF_FFFF value to signify a fully set MASK (i.e., no bits marked
out).

This would mean that if the controller would send an OXe containing
tcp_dst_port_masked:1234/0xFFFF, the switch would store and report
back tcp_dst:1234 (unmasked), which would lead to VFT consistency
errors.