LOXI-83: normalize uint{8,16,32}_t values in ctors
We represent uint8,16,32 values by the next larger numeric type in
Java, i.e., uint8_t is represented by short, not byte. This is done
because the Java types are internally signed and we want unsigned values
for output and input.

However, this creates a problem if somebody inputs a value that is
partially represented by the underlying value, e.g., if they put in a
short "-1" to mean 0xFF. This is a fairly common practice, esp. in C.

Until now this could mean you could have two different OF objects (one
that has short '255' and one that has short '-1' in it). But because
the value uint_8 serializes to 8 bits on the wire, they would come out
to the same on the wire (and for the switch). Two OF objects should be
the same if they are the same on the wire.

Thus, we now normalize those unsigned values to their effective value
range in the constructor of the objects. This ensures that two logically
identical objects (which will be the same on the wire) are also the same
for Java.